Moved OpenEye-dependent schema to receptors module
Also includes various CLI doc improvements
