+++ fah_xchem/schema.py
@@ -3,36 +3,12 @@ import datetime as dt
 from typing import Dict, List, Optional, Union
 
 from pydantic import BaseModel, Field
-from openeye import oechem
 
 
 class Model(BaseModel):
     class Config:
         allow_mutation = False
         extra = "forbid"
-
-## prepare
-
-class OutputPaths(Model):
-    receptor_gzipped: pathlib.Path
-    receptor_thiolate_gzipped: pathlib.Path
-    design_unit_gzipped: pathlib.Path
-    design_unit_thiolate_gzipped: pathlib.Path
-    protein_pdb: pathlib.Path
-    protein_thiolate_pdb: pathlib.Path
-    ligand_pdb: pathlib.Path
-    ligand_sdf: pathlib.Path
-    ligand_mol2: pathlib.Path
-
-
-class DockingSystem(Model):
-    class Config:
-        arbitrary_types_allowed = True
-
-    protein: oechem.OEGraphMol
-    ligand: oechem.OEGraphMol
-    receptor: oechem.OEGraphMol
-    design_unit: oechem.OEDesignUnit
 
 ## analyze and generate
